fixed stdint.h, added varargs

+++ libc/inc/stddef.h
@@ -1,12 +1,10 @@
 #ifndef STDDEF_H
 #define STDDEF_H
 
-#include <stdint.h>
-
 #define NULL ((void*) 0)
 
-typedef uintptr_t size_t;
-typedef uintptr_t ptrdiff_t;
-typedef int32_t wchar_t;
+typedef unsigned int size_t;
+typedef unsigned int ptrdiff_t;
+typedef short   wchar_t;
 
 #endif/*STDDEF_H*/
